move Filters to new view
