+++ hackertracker/Views/EventsView.swift
@@ -89,46 +89,6 @@ struct EventsView: View {
     }
 }
 
-struct EventFilters: View {
-    let tagtypes: [TagType]
-    // let types: [Int: EventType]
-    @Binding var showFilters: Bool
-    @Binding var filters: Set<Int>
-    
-    let gridItemLayout = [GridItem(.flexible()), GridItem(.flexible())]
-
-    var body: some View {
-        NavigationStack {
-            ScrollView {
-                FilterRow(id: 1337, name: "Bookmarks", color: ThemeColors.blue, filters: $filters)
-
-                ForEach(tagtypes.sorted { $0.sortOrder < $1.sortOrder }) { tagtype in
-                    Section(header: Text(tagtype.label)) {
-                        LazyVGrid(columns: gridItemLayout, alignment: .center, spacing: 10) {
-                            ForEach(tagtype.tags.sorted { $0.sortOrder < $1.sortOrder }) { tag in
-                                FilterRow(id: tag.id, name: tag.label, color: Color(UIColor(hex: tag.colorBackground ?? "#2c8f07") ?? .purple), filters: $filters)
-                            }
-                        }
-                    }
-                }.headerProminence(.increased)
-            }
-            .listStyle(.plain)
-            .navigationTitle("Filters")
-            .toolbar {
-                ToolbarItemGroup(placement: .navigationBarTrailing) {
-                    Button("Clear") {
-                        filters.removeAll()
-                    }
-                    Button("Close") {
-                        showFilters = false
-                    }
-                }
-            }
-            .padding(5)
-        }
-    }
-}
-
 struct EventScrollView: View {
     let events: [Date: [Event]]
     let bookmarks: [Int32]
@@ -183,55 +143,5 @@ struct EventData: View {
             }
         }
         .headerProminence(.increased)
-    }
-}
-
-struct FilterRow: View {
-    let id: Int
-    let name: String
-    let color: Color
-    @Binding var filters: Set<Int>
-
-    var body: some View {
-        if !filters.contains(id) {
-            VStack(alignment: .leading) {
-                HStack {
-                    Text(name)
-                        .font(.subheadline)
-                        .padding(5)
-                }
-            }
-            .foregroundColor(.white)
-            .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .leading)
-            .padding(5)
-            .overlay(RoundedRectangle(cornerRadius: 10).stroke(color, lineWidth: 2))
-            .onTapGesture {
-                if filters.contains(id) {
-                    filters.remove(id)
-                } else {
-                    filters.insert(id)
-                }
-            }
-        } else {
-            VStack(alignment: .leading) {
-                HStack {
-                    Text(name)
-                        .font(.subheadline)
-                        .padding(5)
-                }
-            }
-            .foregroundColor(.white)
-            .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .leading)
-            .padding(5)
-            .background(color)
-            .cornerRadius(10)
-            .onTapGesture {
-                if filters.contains(id) {
-                    filters.remove(id)
-                } else {
-                    filters.insert(id)
-                }
-            }
-        }
     }
 }
